+++ opentelemetry-sdk/CHANGELOG.md
@@ -10,7 +10,7 @@
   ([#775](https://github.com/open-telemetry/opentelemetry-python/pull/775))
 - Rename Observer to ValueObserver
   ([#764](https://github.com/open-telemetry/opentelemetry-python/pull/764))
-- Add SumObserver in metrics
+- Add SumObserver and UpDownSumObserver in metrics
   ([#789](https://github.com/open-telemetry/opentelemetry-python/pull/789))
 
 ## 0.8b0